+++ src/spec/util/TestUtils.ts
@@ -8,6 +8,7 @@ export namespace TestUtils {
         width: 500,
         height: 500,
         suppressConsoleBootMessage: true,
+        enableCanvasTransparency: true,
         suppressMinimumBrowserFeatureDetection: true,
         suppressHiDPIScaling: true,
         suppressPlayButton: true,