+++ CHANGELOG.md
@@ -13,6 +13,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
   <!--- Added features here --->
 
+- Added `enableCanvasTransparency` property that will can enable/disable canvas transparency [#1096](https://github.com/excaliburjs/Excalibur/issues/1096)
+
 ### Changed
 
   <!--- Changed behavior here --->